refactored docstrings of kriging module
