+++ pyinterpolate/kriging/models/simple_kriging.py
@@ -1,3 +1,4 @@
+# TODO: log k, predicted, dataset
 # Python core
 from typing import List, Union, Tuple
