+++ pyinterpolate/kriging/utils/process.py
@@ -4,7 +4,7 @@ from typing import List, Union, Tuple
 import numpy as np
 
 from pyinterpolate.distance.distance import calc_point_to_point_distance
-from pyinterpolate.kriging.utils.exceptions import ZerosMatrixWarning, LeastSquaresApproximationWarning
+from pyinterpolate.kriging.utils.kwarnings import ZerosMatrixWarning, LeastSquaresApproximationWarning
 from pyinterpolate.processing.select_values import select_kriging_data
 from pyinterpolate.variogram import TheoreticalVariogram
 from pyinterpolate.variogram.utils.exceptions import validate_theoretical_variogram
